+++ packages/web/src/common/store/cache/collections/sagas.js
@@ -20,8 +20,8 @@ import { getCollection } from 'common/store/cache/collections/selectors'
 import { getTrack } from 'common/store/cache/tracks/selectors'
 import { fetchUsers } from 'common/store/cache/users/sagas'
 import { getUser } from 'common/store/cache/users/selectors'
+import * as signOnActions from 'common/store/pages/signon/actions'
 import { squashNewLines } from 'common/utils/formatUtil'
-import * as signOnActions from 'pages/sign-on/store/actions'
 import { make } from 'store/analytics/actions'
 import * as confirmerActions from 'store/confirmer/actions'
 import { confirmTransaction } from 'store/confirmer/sagas'
