+++ packages/web/src/common/store/cache/tracks/sagas.js
@@ -29,8 +29,8 @@ import * as trackActions from 'common/store/cache/tracks/actions'
 import { getTrack } from 'common/store/cache/tracks/selectors'
 import { fetchUsers } from 'common/store/cache/users/sagas'
 import { getUser } from 'common/store/cache/users/selectors'
+import * as signOnActions from 'common/store/pages/signon/actions'
 import { squashNewLines, formatUrlName } from 'common/utils/formatUtil'
-import * as signOnActions from 'pages/sign-on/store/actions'
 import { fetchCID } from 'services/audius-backend'
 import TrackDownload from 'services/audius-backend/TrackDownload'
 import { make } from 'store/analytics/actions'
