+++ packages/web/src/common/store/pages/signon/sagas.js
@@ -33,7 +33,6 @@ import { saveCollection } from 'common/store/social/collections/actions'
 import * as socialActions from 'common/store/social/users/actions'
 import { getFeePayer } from 'common/store/solana/selectors'
 import { ELECTRONIC_SUBGENRES, Genre } from 'common/utils/genres'
-import { getIGUserUrl } from 'components/instagram-auth/InstagramAuth'
 import { getCityAndRegion } from 'services/Location'
 import { identify, make } from 'store/analytics/actions'
 import * as confirmerActions from 'store/confirmer/actions'
@@ -45,7 +44,7 @@ import { restrictedHandles } from 'utils/restrictedHandles'
 import { ERROR_PAGE, FEED_PAGE, SIGN_IN_PAGE, SIGN_UP_PAGE } from 'utils/route'
 import { waitForAccount } from 'utils/sagaHelpers'
 
-import { MAX_HANDLE_LENGTH } from '../utils/formatSocialProfile'
+import { MAX_HANDLE_LENGTH } from '../../../../pages/sign-on/utils/formatSocialProfile'
 
 import * as signOnActions from './actions'
 import { watchSignOnError } from './errorSagas'
@@ -62,6 +61,12 @@ const NATIVE_MOBILE = process.env.REACT_APP_NATIVE_MOBILE
 const SUGGESTED_FOLLOW_USER_HANDLE_URL =
   process.env.REACT_APP_SUGGESTED_FOLLOW_HANDLES
 const SIGN_UP_TIMEOUT_MILLIS = 20 /* min */ * 60 * 1000
+
+// Route to fetch instagram user data w/ the username
+export const getIGUserUrl = (endpoint: string, username: string) => {
+  const url = endpoint.replace('$USERNAME$', username)
+  return url
+}
 
 const messages = {
   incompleteAccount:
@@ -434,6 +439,7 @@ function* signIn(action) {
   yield call(waitForBackendSetup)
   try {
     const signOn = yield select(getSignOn)
+    console.log({ signOn })
     const signInResponse = yield call(
       audiusBackendInstance.signIn,
       signOn.email.value,
