+++ packages/web/src/components/download-buttons/DownloadButtons.tsx
@@ -11,15 +11,15 @@ import {
   ButtonType,
   useDownloadTrackButtons
 } from 'common/hooks/useDownloadTrackButtons'
-import { toast } from 'common/store/ui/toast/slice'
-import LoadingSpinner from 'components/loading-spinner/LoadingSpinner'
-import Tooltip from 'components/tooltip/Tooltip'
 import {
   openSignOn,
   updateRouteOnExit,
   updateRouteOnCompletion,
   showRequiresAccountModal
-} from 'pages/sign-on/store/actions'
+} from 'common/store/pages/signon/actions'
+import { toast } from 'common/store/ui/toast/slice'
+import LoadingSpinner from 'components/loading-spinner/LoadingSpinner'
+import Tooltip from 'components/tooltip/Tooltip'
 import { useIsMobile } from 'utils/clientUtil'
 
 import styles from './DownloadButtons.module.css'
