+++ packages/web/src/components/nav/desktop/NavColumn.tsx
@@ -42,6 +42,7 @@ import {
   getNotificationPanelIsOpen,
   getNotificationUnviewedCount
 } from 'common/store/notifications/selectors'
+import * as signOnActions from 'common/store/pages/signon/actions'
 import {
   addFolderToLibrary,
   constructPlaylistFolder
@@ -67,7 +68,6 @@ import Tooltip from 'components/tooltip/Tooltip'
 import UserBadges from 'components/user-badges/UserBadges'
 import { useFlag } from 'hooks/useRemoteConfig'
 import { useUserProfilePicture } from 'hooks/useUserProfilePicture'
-import * as signOnActions from 'pages/sign-on/store/actions'
 import { resetState as resetUploadState } from 'pages/upload-page/store/actions'
 import { NO_VISUALIZER_ROUTES } from 'pages/visualizer/Visualizer'
 import { openVisualizer } from 'pages/visualizer/store/slice'
