+++ packages/web/src/components/nav/mobile/ConnectedNavBar.tsx
@@ -12,11 +12,11 @@ import {
 } from 'common/store/account/selectors'
 import { getNotificationUnviewedCount } from 'common/store/notifications/selectors'
 import { getSearchStatus } from 'common/store/pages/search-results/selectors'
+import { openSignOn } from 'common/store/pages/signon/actions'
 import {
   RouterContext,
   SlideDirection
 } from 'components/animated-switch/RouterContextProvider'
-import { openSignOn } from 'pages/sign-on/store/actions'
 import { make, useRecord } from 'store/analytics/actions'
 import { AppState } from 'store/types'
 import { getIsIOS } from 'utils/browser'
