+++ packages/web/src/hooks/useRequiresAccount.ts
@@ -7,7 +7,7 @@ import { getAccountUser } from 'common/store/account/selectors'
 import {
   showRequiresAccountModal,
   updateRouteOnExit
-} from 'pages/sign-on/store/actions'
+} from 'common/store/pages/signon/actions'
 import { useSelector } from 'utils/reducer'
 import { SIGN_UP_PAGE } from 'utils/route'
