+++ packages/web/src/pages/feed-page/store/lineups/feed/sagas.ts
@@ -17,12 +17,12 @@ import { processAndCacheCollections } from 'common/store/cache/collections/utils
 import { processAndCacheTracks } from 'common/store/cache/tracks/utils'
 import { PREFIX, feedActions } from 'common/store/pages/feed/lineup/actions'
 import { getFeedFilter } from 'common/store/pages/feed/selectors'
-import { CommonState } from 'common/store/reducers'
 import {
   getAccountReady,
   getFollowIds,
   getStartedSignOnProcess
-} from 'pages/sign-on/store/selectors'
+} from 'common/store/pages/signon/selectors'
+import { CommonState } from 'common/store/reducers'
 import { LineupSagas } from 'store/lineup/sagas'
 import { waitForAccount } from 'utils/sagaHelpers'
