+++ packages/web/src/pages/sign-on/components/desktop/FollowPage.tsx
@@ -8,7 +8,10 @@ import { ReactComponent as IconWand } from 'assets/img/iconWand.svg'
 import UserCard from 'components/card/UserCard'
 import SelectablePills from 'components/selectable-pill/SelectablePills'
 
-import { FollowArtistsCategory, artistCategories } from '../../store/types'
+import {
+  FollowArtistsCategory,
+  artistCategories
+} from '../../../../common/store/pages/signon/types'
 
 import styles from './FollowPage.module.css'
