+++ packages/web/src/pages/sign-on/components/desktop/SignOnPage.tsx
@@ -17,6 +17,8 @@ import {
   TwitterProfile
 } from 'common/store/account/reducer'
 import { getAccountStatus } from 'common/store/account/selectors'
+import { getStatus } from 'common/store/pages/signon/selectors'
+import { Pages, FollowArtistsCategory } from 'common/store/pages/signon/types'
 import BackgroundWaves from 'components/background-animations/BackgroundWaves'
 import Page from 'components/page/Page'
 import SignOnModal from 'components/sign-on/SignOnModal'
@@ -29,8 +31,6 @@ import PasswordPage from 'pages/sign-on/components/desktop/PasswordPage'
 import ProfilePage from 'pages/sign-on/components/desktop/ProfilePage'
 import { SignInPage } from 'pages/sign-on/components/desktop/SignInPage'
 import StartPlatformPage from 'pages/sign-on/components/desktop/StartPlatformPage'
-import { getStatus } from 'pages/sign-on/store/selectors'
-import { Pages, FollowArtistsCategory } from 'pages/sign-on/store/types'
 import lazyWithPreload from 'utils/lazyWithPreload'
 import { useSelector } from 'utils/reducer'
 import { BASE_URL, SIGN_UP_PAGE } from 'utils/route'
