+++ packages/web/src/pages/sign-on/components/mobile/FollowPage.tsx
@@ -9,7 +9,10 @@ import UserCard from 'components/card/UserCard'
 import SelectablePills from 'components/selectable-pill/SelectablePills'
 import { MAIN_CONTENT_ID } from 'pages/App'
 
-import { FollowArtistsCategory, artistCategories } from '../../store/types'
+import {
+  FollowArtistsCategory,
+  artistCategories
+} from '../../../../common/store/pages/signon/types'
 
 import styles from './FollowPage.module.css'
