+++ packages/web/src/pages/trending-page/TrendingPageProvider.js
@@ -10,6 +10,7 @@ import { matchPath, withRouter } from 'react-router-dom'
 
 import { getHasAccount } from 'common/store/account/selectors'
 import { makeGetLineupMetadatas } from 'common/store/lineup/selectors'
+import { openSignOn } from 'common/store/pages/signon/actions'
 import * as trendingPageActions from 'common/store/pages/trending/actions'
 import {
   trendingActions,
@@ -27,7 +28,6 @@ import {
 } from 'common/store/pages/trending/selectors'
 import { makeGetCurrent } from 'common/store/queue/selectors'
 import { GENRES } from 'common/utils/genres'
-import { openSignOn } from 'pages/sign-on/store/actions'
 import { make } from 'store/analytics/actions'
 import { getBuffering, getPlaying } from 'store/player/selectors'
 import { isMobile } from 'utils/clientUtil'
