+++ packages/web/src/store/social/collections/sagas.ts
@@ -22,10 +22,10 @@ import {
 import { adjustUserField } from 'common/store/cache/users/sagas'
 import { getUser } from 'common/store/cache/users/selectors'
 import * as notificationActions from 'common/store/notifications/actions'
+import * as signOnActions from 'common/store/pages/signon/actions'
 import { removeFromPlaylistLibrary } from 'common/store/playlist-library/helpers'
 import * as socialActions from 'common/store/social/collections/actions'
 import { formatShareText } from 'common/utils/formatUtil'
-import * as signOnActions from 'pages/sign-on/store/actions'
 import { make } from 'store/analytics/actions'
 import * as confirmerActions from 'store/confirmer/actions'
 import { confirmTransaction } from 'store/confirmer/sagas'
