+++ packages/web/src/store/social/tracks/sagas.ts
@@ -10,9 +10,9 @@ import { getTrack, getTracks } from 'common/store/cache/tracks/selectors'
 import { adjustUserField } from 'common/store/cache/users/sagas'
 import { getUser } from 'common/store/cache/users/selectors'
 import { updateOptimisticListenStreak } from 'common/store/pages/audio-rewards/slice'
+import * as signOnActions from 'common/store/pages/signon/actions'
 import * as socialActions from 'common/store/social/tracks/actions'
 import { formatShareText } from 'common/utils/formatUtil'
-import * as signOnActions from 'pages/sign-on/store/actions'
 import TrackDownload from 'services/audius-backend/TrackDownload'
 import { make } from 'store/analytics/actions'
 import * as confirmerActions from 'store/confirmer/actions'
