+++ packages/web/src/store/social/users/sagas.ts
@@ -7,8 +7,8 @@ import { waitForBackendSetup } from 'common/store/backend/sagas'
 import * as cacheActions from 'common/store/cache/actions'
 import { adjustUserField } from 'common/store/cache/users/sagas'
 import { getUsers, getUser } from 'common/store/cache/users/selectors'
+import * as signOnActions from 'common/store/pages/signon/actions'
 import * as socialActions from 'common/store/social/users/actions'
-import * as signOnActions from 'pages/sign-on/store/actions'
 import { make } from 'store/analytics/actions'
 import * as confirmerActions from 'store/confirmer/actions'
 import { confirmTransaction } from 'store/confirmer/sagas'
