+++ packages/web/src/store/storeContext.ts
@@ -17,5 +17,6 @@ export const storeContext: CommonStoreContext = {
   apiClient,
   fingerprintClient,
   walletClient,
-  localStorage
+  localStorage,
+  isNativeMobile: false
 }
