+++ packages/web/src/store/types.ts
@@ -7,6 +7,7 @@ import NotificationState from 'common/store/notifications/types'
 import { CollectionsPageState } from 'common/store/pages/collection/types'
 import HistoryPageState from 'common/store/pages/history-page/types'
 import RemixesPageReducer from 'common/store/pages/remixes/slice'
+import SignOnPageState from 'common/store/pages/signon/types'
 import { SmartCollectionState } from 'common/store/pages/smart-collection/slice'
 import QueueReducer from 'common/store/queue/slice'
 import { ReachabilityState } from 'common/store/reachability/types'
@@ -29,7 +30,6 @@ import { UnfollowConfirmationModalState } from 'components/unfollow-confirmation
 import ArtistDashboardState from 'pages/artist-dashboard-page/store/types'
 import { DeactivateAccountState } from 'pages/deactivate-account-page/store/slice'
 import DeletedPageReducer from 'pages/deleted-page/store/slice'
-import SignOnPageState from 'pages/sign-on/store/types'
 import { UploadPageState } from 'pages/upload-page/store/types'
 import VisualizerReducer from 'pages/visualizer/store/slice'
 import AppCTAModalReducer from 'store/application/ui/app-cta-modal/slice'
